# contracts/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Shifa Platform API",
    "description": "High-performance modular API for Shifa WhatsApp pharmacy commerce platform in Pakistan.",
    "license": {
      "name": "Proprietary"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokens"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "logout",
        "responses": {
          "200": {
            "description": "Logged out successfully"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "operationId": "me",
        "responses": {
          "200": {
            "description": "Current user profile and permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/auth/password/change": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "change_password",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Password changed successfully"
          },
          "400": {
            "description": "Password validation error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "refresh",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Tokens refreshed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokens"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or revoked refresh token"
          }
        }
      }
    },
    "/api/v1/branches": {
      "get": {
        "tags": [
          "Branches"
        ],
        "operationId": "list_branches",
        "responses": {
          "200": {
            "description": "List branches",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BranchDto"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Branches"
        ],
        "operationId": "create_branch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBranchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Branch created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BranchDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/branches/{id}": {
      "patch": {
        "tags": [
          "Branches"
        ],
        "operationId": "update_branch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Branch ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBranchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Branch updated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Branch not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/permissions": {
      "get": {
        "tags": [
          "Roles"
        ],
        "operationId": "list_permissions",
        "responses": {
          "200": {
            "description": "List all permissions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/roles": {
      "get": {
        "tags": [
          "Roles"
        ],
        "operationId": "list_roles",
        "responses": {
          "200": {
            "description": "List roles",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoleDto"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "operationId": "list_users",
        "responses": {
          "200": {
            "description": "List users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDto"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Users"
        ],
        "operationId": "create_user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/users/{id}": {
      "delete": {
        "tags": [
          "Users"
        ],
        "operationId": "delete_user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Target user ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User soft deleted"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Users"
        ],
        "operationId": "update_user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Target user ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "User updated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/users/{id}/branches": {
      "post": {
        "tags": [
          "Users"
        ],
        "operationId": "assign_branches",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Target user ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignBranchesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Branches assigned"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/users/{id}/roles": {
      "post": {
        "tags": [
          "Users"
        ],
        "operationId": "assign_roles",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Target user ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRolesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Roles assigned"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "AssignBranchesRequest": {
        "type": "object",
        "required": [
          "branch_ids"
        ],
        "properties": {
          "branch_ids": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BranchId"
            }
          }
        }
      },
      "AssignRolesRequest": {
        "type": "object",
        "required": [
          "role_names"
        ],
        "properties": {
          "role_names": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "AuthTokens": {
        "type": "object",
        "required": [
          "access_token",
          "refresh_token",
          "token_type",
          "expires_in"
        ],
        "properties": {
          "access_token": {
            "type": "string"
          },
          "expires_in": {
            "type": "integer",
            "format": "int64"
          },
          "refresh_token": {
            "type": "string"
          },
          "token_type": {
            "type": "string"
          }
        }
      },
      "BranchDto": {
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "name",
          "code",
          "drap_licence_no",
          "pharmacist_in_charge",
          "address",
          "city",
          "is_hub",
          "cold_chain_capable",
          "status"
        ],
        "properties": {
          "address": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "cold_chain_capable": {
            "type": "boolean"
          },
          "drap_licence_no": {
            "type": "string"
          },
          "id": {
            "$ref": "#/components/schemas/BranchId"
          },
          "is_hub": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "pharmacist_in_charge": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "tenant_id": {
            "$ref": "#/components/schemas/TenantId"
          }
        }
      },
      "BranchId": {
        "type": "string",
        "format": "uuid",
        "example": "018f3a9e-4c5b-7b3a-9e1a-2b3c4d5e6f7a"
      },
      "ChangePasswordRequest": {
        "type": "object",
        "required": [
          "current_password",
          "new_password"
        ],
        "properties": {
          "current_password": {
            "type": "string"
          },
          "new_password": {
            "type": "string"
          }
        }
      },
      "CreateBranchRequest": {
        "type": "object",
        "required": [
          "name",
          "code",
          "drap_licence_no",
          "pharmacist_in_charge",
          "address",
          "city",
          "longitude",
          "latitude",
          "is_hub",
          "cold_chain_capable"
        ],
        "properties": {
          "address": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "cold_chain_capable": {
            "type": "boolean"
          },
          "drap_licence_no": {
            "type": "string"
          },
          "is_hub": {
            "type": "boolean"
          },
          "latitude": {
            "type": "number",
            "format": "double"
          },
          "longitude": {
            "type": "number",
            "format": "double"
          },
          "name": {
            "type": "string"
          },
          "pharmacist_in_charge": {
            "type": "string"
          }
        }
      },
      "CreateUserRequest": {
        "type": "object",
        "required": [
          "phone",
          "full_name",
          "password",
          "role_names",
          "branch_ids"
        ],
        "properties": {
          "branch_ids": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BranchId"
            }
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "full_name": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "role_names": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": [
          "phone_or_email",
          "password"
        ],
        "properties": {
          "password": {
            "type": "string"
          },
          "phone_or_email": {
            "type": "string"
          }
        }
      },
      "RefreshRequest": {
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "properties": {
          "refresh_token": {
            "type": "string"
          }
        }
      },
      "RoleDto": {
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "name",
          "is_system",
          "permissions"
        ],
        "properties": {
          "description": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "$ref": "#/components/schemas/RoleId"
          },
          "is_system": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "tenant_id": {
            "$ref": "#/components/schemas/TenantId"
          }
        }
      },
      "RoleId": {
        "type": "string",
        "format": "uuid",
        "example": "018f3a9e-4c5b-7b3a-9e1a-2b3c4d5e6f7a"
      },
      "TenantId": {
        "type": "string",
        "format": "uuid",
        "example": "018f3a9e-4c5b-7b3a-9e1a-2b3c4d5e6f7a"
      },
      "UpdateBranchRequest": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string",
            "nullable": true
          },
          "cold_chain_capable": {
            "type": "boolean",
            "nullable": true
          },
          "is_hub": {
            "type": "boolean",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "pharmacist_in_charge": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "UpdateUserRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "full_name": {
            "type": "string",
            "nullable": true
          },
          "locale": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "UserDto": {
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "phone",
          "full_name",
          "status",
          "locale",
          "roles",
          "branch_ids",
          "created_at"
        ],
        "properties": {
          "branch_ids": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BranchId"
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "full_name": {
            "type": "string"
          },
          "id": {
            "$ref": "#/components/schemas/UserId"
          },
          "last_login_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "locale": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string"
          },
          "tenant_id": {
            "$ref": "#/components/schemas/TenantId"
          }
        }
      },
      "UserId": {
        "type": "string",
        "format": "uuid",
        "example": "018f3a9e-4c5b-7b3a-9e1a-2b3c4d5e6f7a"
      },
      "UserProfileResponse": {
        "type": "object",
        "required": [
          "user",
          "permissions"
        ],
        "properties": {
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "user": {
            "$ref": "#/components/schemas/UserDto"
          }
        }
      }
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "tags": [
    {
      "name": "Auth",
      "description": "Authentication and session management"
    },
    {
      "name": "Users",
      "description": "User management and RBAC assignments"
    },
    {
      "name": "Branches",
      "description": "Branch store locations and configuration"
    },
    {
      "name": "Roles",
      "description": "Roles and permissions directory"
    }
  ]
}